+++ src/gui/element/my_addons.rs
@@ -1,8 +1,8 @@
 use {
     super::{DEFAULT_FONT_SIZE, DEFAULT_PADDING},
     crate::gui::{
-        style, ColumnKey, ColumnState, ExpandType, Flavor, Interaction, Message, Mode,
-        ReleaseChannel, SortDirection, State,
+        style, AddonsSearchState, ColumnKey, ColumnState, ExpandType, Flavor, Interaction, Message,
+        Mode, ReleaseChannel, SortDirection, State,
     },
     crate::localization::{localized_string, localized_timeago_formatter},
     ajour_core::{
@@ -12,7 +12,10 @@ use {
     },
     ajour_widgets::{header, Header, TableRow},
     chrono::prelude::*,
-    iced::{button, Align, Button, Column, Container, Element, Length, PickList, Row, Space, Text},
+    iced::{
+        button, Align, Button, Column, Container, Element, Length, PickList, Row, Space, Text,
+        TextInput,
+    },
     std::collections::HashMap,
     strfmt::strfmt,
 };
@@ -742,6 +745,7 @@ pub fn menu_container<'a>(
     flavor: Flavor,
     update_all_button_state: &'a mut button::State,
     refresh_button_state: &'a mut button::State,
+    addons_search_state: &'a mut AddonsSearchState,
     state: &HashMap<Mode, State>,
     addons: &[Addon],
     config: &Config,
@@ -753,7 +757,7 @@ pub fn menu_container<'a>(
         .unwrap_or_default();
 
     // A row contain general settings.
-    let mut settings_row = Row::new().height(Length::Units(35));
+    let mut settings_row = Row::new().align_items(Align::Center);
 
     let mut update_all_button = Button::new(
         update_all_button_state,
@@ -818,6 +822,20 @@ pub fn menu_container<'a>(
         _ => Text::new(""),
     };
 
+    let query = addons_search_state.query.as_deref().unwrap_or_default();
+    let addons_query = TextInput::new(
+        &mut addons_search_state.query_state,
+        &localized_string("search-for-addon")[..],
+        query,
+        Interaction::AddonsQuery,
+    )
+    .size(DEFAULT_FONT_SIZE)
+    .padding(7)
+    .width(Length::Units(225))
+    .style(style::AddonsQueryInput(color_palette));
+
+    let addons_query: Element<Interaction> = addons_query.into();
+
     let status_container = Container::new(status_text)
         .center_y()
         .padding(5)
@@ -831,12 +849,18 @@ pub fn menu_container<'a>(
         .push(update_all_button.map(Message::Interaction))
         .push(Space::new(Length::Units(7), Length::Units(0)))
         .push(status_container)
-        .push(Space::new(Length::Units(DEFAULT_PADDING), Length::Units(0)));
+        .push(Space::new(Length::Fill, Length::Units(0)))
+        .push(addons_query.map(Message::Interaction))
+        .push(Space::new(
+            Length::Units(DEFAULT_PADDING + 5),
+            Length::Units(0),
+        ));
 
     // Add space above settings_row.
     let settings_column = Column::new()
         .push(Space::new(Length::Units(0), Length::Units(5)))
-        .push(settings_row);
+        .push(settings_row)
+        .push(Space::new(Length::Units(0), Length::Units(8)));
 
     // Wraps it in a container.
     Container::new(settings_column)